+++ src/plugins/intel_gpu/src/graph/network.cpp
@@ -541,7 +541,7 @@ void network::save(cldnn::BinaryOutputBuffer& ob) {
     }
 
     int exec_order_size;
-    exec_order_size = _exec_order.size();
+    exec_order_size = static_cast<int>(_exec_order.size());
     ob << exec_order_size;
 
     for (const auto& p_inst : _exec_order) {
